+++ payments/payment_gateways/doctype/razorpay_settings/test_razorpay_settings.py
@@ -1,8 +1,5 @@
 # Copyright (c) 2026, Frappe Technologies and Contributors
 # License: MIT. See LICENSE
-import hashlib
-import hmac
-import json
 from unittest.mock import MagicMock, patch
 
 import frappe
@@ -12,9 +9,6 @@ from frappe.tests import IntegrationTestCase, UnitTestCase
 from payments.payment_gateways.doctype.razorpay_settings.razorpay_settings import (
 	RazorpaySettings,
 	from_paise,
-	handle_refund_notification,
-	process_webhook,
-	razorpay_webhook,
 	to_paise,
 )
 
@@ -125,144 +119,3 @@ class TestRazorpayRefund(IntegrationTestCase):
 				self.settings.refund_payment("pay_123", 10)
 
 		self.assertIn("atleast INR 1.00", str(raised.exception))
-
-
-REFUND_PROCESSED_PAYLOAD = {
-	"event": "refund.processed",
-	"payload": {
-		"refund": {
-			"entity": {
-				"id": "rfnd_1",
-				"status": "processed",
-				"amount": 6000,
-				"payment_id": "pay_123",
-			}
-		}
-	},
-}
-
-
-def sign(body: bytes, secret: str) -> str:
-	return hmac.new(secret.encode(), body, hashlib.sha256).hexdigest()
-
-
-class TestRazorpayWebhook(IntegrationTestCase):
-	def setUp(self):
-		self.secret = "whsec_test"
-		self.body = json.dumps(REFUND_PROCESSED_PAYLOAD).encode()
-		self.patcher = patch.object(RazorpaySettings, "get_password", return_value=self.secret)
-		self.patcher.start()
-		self.addCleanup(self.patcher.stop)
-
-	def test_rejects_a_bad_signature(self):
-		self.assertRaises(frappe.PermissionError, process_webhook, self.body, "deadbeef")
-
-	def test_rejects_a_webhook_when_no_secret_is_configured(self):
-		# Anyone can compute an HMAC under an empty key, forged payloads included.
-		self.patcher.stop()
-		for unset in (None, ""):
-			with self.subTest(secret=unset):
-				with patch.object(RazorpaySettings, "get_password", return_value=unset):
-					self.assertRaises(
-						frappe.ValidationError, process_webhook, self.body, sign(self.body, unset or "")
-					)
-		self.patcher.start()
-
-	def test_ignores_unsupported_events(self):
-		body = json.dumps(dict(REFUND_PROCESSED_PAYLOAD, event="payment.captured")).encode()
-
-		self.assertIsNone(process_webhook(body, sign(body, self.secret)))
-
-	def test_logs_a_supported_event_as_an_integration_request(self):
-		name = process_webhook(self.body, sign(self.body, self.secret))
-
-		log = frappe.get_doc("Integration Request", name)
-		self.assertEqual(log.status, "Queued")
-		self.assertEqual(json.loads(log.data)["event"], "refund.processed")
-
-	def test_a_rejected_webhook_logs_nothing(self):
-		before = frappe.db.count("Integration Request")
-
-		self.assertRaises(frappe.PermissionError, process_webhook, self.body, "deadbeef")
-
-		self.assertEqual(frappe.db.count("Integration Request"), before)
-
-
-class TestRazorpayWebhookEndpoint(IntegrationTestCase):
-	def setUp(self):
-		self.secret = "whsec_test"
-		self.body = json.dumps(REFUND_PROCESSED_PAYLOAD).encode()
-		patcher = patch.object(RazorpaySettings, "get_password", return_value=self.secret)
-		patcher.start()
-		self.addCleanup(patcher.stop)
-
-	def post(self, signature):
-		request = MagicMock()
-		request.data = self.body
-
-		with (
-			patch.object(frappe, "request", request),
-			patch.object(frappe, "get_request_header", return_value=signature),
-			patch.object(frappe, "enqueue") as enqueue,
-		):
-			return razorpay_webhook(), enqueue
-
-	def test_a_bad_signature_is_answered_normally_and_left_in_the_error_log(self):
-		integration_requests = frappe.db.count("Integration Request")
-		error_logs = frappe.db.count("Error Log")
-
-		response, enqueue = self.post("deadbeef")
-
-		self.assertIsNone(response)
-		enqueue.assert_not_called()
-		self.assertEqual(frappe.db.count("Integration Request"), integration_requests)
-		self.assertEqual(frappe.db.count("Error Log"), error_logs + 1)
-
-	def test_a_valid_webhook_is_queued_for_processing(self):
-		_, enqueue = self.post(sign(self.body, self.secret))
-
-		self.assertEqual(enqueue.call_args.kwargs["doctype"], "Integration Request")
-		self.assertEqual(
-			frappe.db.get_value("Integration Request", enqueue.call_args.kwargs["docname"], "status"),
-			"Queued",
-		)
-
-
-class TestRazorpayRefundNotification(IntegrationTestCase):
-	def setUp(self):
-		self.log = frappe.get_doc(
-			{
-				"doctype": "Integration Request",
-				"integration_request_service": "Razorpay",
-				"request_description": "Refund Notification",
-				"data": json.dumps(REFUND_PROCESSED_PAYLOAD),
-				"is_remote_request": 1,
-				"status": "Queued",
-			}
-		).insert(ignore_permissions=True)
-
-	def test_a_handled_notification_completes_the_request(self):
-		with patch(
-			"payments.payment_gateways.doctype.razorpay_settings.razorpay_settings.call_hook_method"
-		) as hook:
-			handle_refund_notification("Integration Request", self.log.name)
-
-		hook.assert_called_once_with(
-			"handle_refund_notification", doctype="Integration Request", docname=self.log.name
-		)
-		self.assertEqual(frappe.db.get_value("Integration Request", self.log.name, "status"), "Completed")
-
-	def test_a_subscriber_failure_is_rolled_back_and_recorded_on_the_request(self):
-		with (
-			patch(
-				"payments.payment_gateways.doctype.razorpay_settings.razorpay_settings.call_hook_method",
-				side_effect=Exception("subscriber blew up"),
-			),
-			patch.object(frappe.db, "rollback") as rollback,
-		):
-			handle_refund_notification("Integration Request", self.log.name)
-
-		rollback.assert_called_once()
-		status, error = frappe.db.get_value("Integration Request", self.log.name, ["status", "error"])
-		self.assertEqual(status, "Failed")
-		self.assertIn("subscriber blew up", error)
